+++ src/matrix.rs
@@ -1,9 +1,8 @@
-use crate::Float;
 
 pub trait Matrix: Sync + Send {
-    fn new(a_len: usize, b_len: usize, init_val: Float) -> Self;
-    fn set_diagonal_cell(&mut self, diag_row: usize, diag_offset: isize, value: Float);
-    fn get_diagonal_cell(&self, diag_row: usize, diag_offset: isize) -> Float;
+    fn new(a_len: usize, b_len: usize, init_val: f64) -> Self;
+    fn set_diagonal_cell(&mut self, diag_row: usize, diag_offset: isize, value: f64);
+    fn get_diagonal_cell(&self, diag_row: usize, diag_offset: isize) -> f64;
 
     fn index_mat_to_diag(i: usize, j: usize) -> (usize, isize) {
         (i + j, (j as isize) - (i as isize))
@@ -16,14 +15,14 @@ pub trait Matrix: Sync + Send {
 }
 
 pub struct FullMatrix {
-    matrix: Vec<Vec<Float>>,
+    matrix: Vec<Vec<f64>>,
     diag_len: usize,
     a_len: usize,
     b_len: usize,
 }
 
 impl Matrix for FullMatrix {
-    fn new(a_len: usize, b_len: usize, init_val: Float) -> Self {
+    fn new(a_len: usize, b_len: usize, init_val: f64) -> Self {
         let rowcount = a_len + b_len + 1;
         let diag_len = a_len + b_len + 1;
 
@@ -35,11 +34,11 @@ impl Matrix for FullMatrix {
         }
     }
 
-    fn get_diagonal_cell(&self, diag_row: usize, diag_offset: isize) -> Float {
+    fn get_diagonal_cell(&self, diag_row: usize, diag_offset: isize) -> f64 {
         self.matrix[diag_row][self.diag_len.overflowing_add_signed(diag_offset).0 % self.diag_len]
     }
 
-    fn set_diagonal_cell(&mut self, diag_row: usize, diag_offset: isize, value: Float) {
+    fn set_diagonal_cell(&mut self, diag_row: usize, diag_offset: isize, value: f64) {
         self.matrix[diag_row]
             [self.diag_len.overflowing_add_signed(diag_offset).0 % self.diag_len] = value;
     }
@@ -67,12 +66,12 @@ impl Matrix for FullMatrix {
 }
 
 pub struct DiagonalMatrix {
-    diagonal: Vec<Float>,
+    diagonal: Vec<f64>,
     mask: usize,
 }
 
 impl Matrix for DiagonalMatrix {
-    fn new(a_len: usize, _b_len: usize, init_val: Float) -> Self {
+    fn new(a_len: usize, _b_len: usize, init_val: f64) -> Self {
         let diag_len = 2 * (a_len + 1).next_power_of_two();
 
         Self {
@@ -82,12 +81,12 @@ impl Matrix for DiagonalMatrix {
     }
 
     #[inline(always)]
-    fn get_diagonal_cell(&self, _diag_row: usize, diag_offset: isize) -> Float {
+    fn get_diagonal_cell(&self, _diag_row: usize, diag_offset: isize) -> f64 {
         self.diagonal[diag_offset as usize & self.mask]
     }
 
     #[inline(always)]
-    fn set_diagonal_cell(&mut self, _diag_row: usize, diag_offset: isize, value: Float) {
+    fn set_diagonal_cell(&mut self, _diag_row: usize, diag_offset: isize, value: f64) {
         self.diagonal[diag_offset as usize & self.mask] = value;
     }
 
@@ -102,14 +101,14 @@ pub struct CheckMatrix {
 }
 
 impl Matrix for CheckMatrix {
-    fn new(a_len: usize, b_len: usize, init_val: Float) -> Self {
+    fn new(a_len: usize, b_len: usize, init_val: f64) -> Self {
         Self {
             full: FullMatrix::new(a_len, b_len, init_val),
             optim: DiagonalMatrix::new(a_len, b_len, init_val),
         }
     }
 
-    fn get_diagonal_cell(&self, diag_row: usize, diag_offset: isize) -> Float {
+    fn get_diagonal_cell(&self, diag_row: usize, diag_offset: isize) -> f64 {
         let full = self.full.get_diagonal_cell(diag_row, diag_offset);
         let optim = self.optim.get_diagonal_cell(diag_row, diag_offset);
         if full != optim {
@@ -125,7 +124,7 @@ impl Matrix for CheckMatrix {
         full
     }
 
-    fn set_diagonal_cell(&mut self, diag_row: usize, diag_offset: isize, value: Float) {
+    fn set_diagonal_cell(&mut self, diag_row: usize, diag_offset: isize, value: f64) {
         // if diag_offset == -14 {
         //     println!("Setting diag_offset: {} to value: {}", diag_offset, value);
         // }
