+++ src/utils.rs
@@ -1,12 +1,10 @@
 use rustfft::{Fft, FftDirection, algorithm::Radix4, num_complex::Complex};
 
-use crate::Float;
-
 pub fn next_multiple_of_n(x: usize, n: usize) -> usize {
     (x + n - 1) / n * n
 }
 
-pub fn derivate(x: &Vec<Vec<Float>>) -> Vec<Vec<Float>> {
+pub fn derivate(x: &Vec<Vec<f64>>) -> Vec<Vec<f64>> {
     let mut x_d = Vec::with_capacity(x.len());
     for i in 0..x.len() {
         x_d.push(vec![0.0; x[i].len() - 2]);
@@ -19,25 +17,25 @@ pub fn derivate(x: &Vec<Vec<Float>>) -> Vec<Vec<Float>> {
     x_d
 }
 
-const WEIGHT_MAX: Float = 1.0;
-pub fn dtw_weights(len: usize, g: Float) -> Vec<Float> {
+const WEIGHT_MAX: f64 = 1.0;
+pub fn dtw_weights(len: usize, g: f64) -> Vec<f64> {
     let mut weights = vec![0.0; len];
-    let half_len = len as Float / 2.0;
+    let half_len = len as f64 / 2.0;
     for i in 0..len {
         weights[i] =
-            WEIGHT_MAX / (1.0 + (std::f64::consts::E as Float).powf(-g * (i as Float - half_len as Float)));
+            WEIGHT_MAX / (1.0 + (std::f64::consts::E as f64).powf(-g * (i as f64 - half_len as f64)));
     }
     weights
 }
 // [1 / (1 + np.exp(-g * (i - max_size / 2))) for i in range(0, max_size)]
 
-const MSM_C: Float = 1.0;
+const MSM_C: f64 = 1.0;
 #[inline(always)]
-pub fn msm_cost_function(x: Float, y: Float, z: Float) -> Float {
+pub fn msm_cost_function(x: f64, y: f64, z: f64) -> f64 {
     MSM_C + (y.min(z) - x).max(x - y.max(z)).max(0.0)
 }
 
-pub fn cross_correlation(a: &[Float], b: &[Float]) -> Vec<Float> {
+pub fn cross_correlation(a: &[f64], b: &[f64]) -> Vec<f64> {
     // zero-pad the input signals a and b (add zeros to the end of each. The zero padding should fill the vectors until they reach a size of at least N = size(a)+size(b)-1
     let fft_len = (a.len() + b.len() - 1).next_power_of_two();
     let fft = Radix4::new(fft_len, FftDirection::Forward);
@@ -63,17 +61,17 @@ pub fn cross_correlation(a: &[Float], b: &[Float]) -> Vec<Float> {
     let ifft = Radix4::new(fft_len, FftDirection::Inverse);
     ifft.process(&mut c_fft);
     for i in 0..fft_len {
-        c[i] = c_fft[i].re / fft_len as Float;
+        c[i] = c_fft[i].re / fft_len as f64;
     }
     c
 }
 
-pub fn zscore(x: &[Float]) -> Vec<Float> {
-    let mean = x.iter().sum::<Float>() / x.len() as Float;
-    let std = (x.iter().map(|val| (val - mean).powi(2)).sum::<Float>() / x.len() as Float).sqrt();
+pub fn zscore(x: &[f64]) -> Vec<f64> {
+    let mean = x.iter().sum::<f64>() / x.len() as f64;
+    let std = (x.iter().map(|val| (val - mean).powi(2)).sum::<f64>() / x.len() as f64).sqrt();
     x.iter().map(|val| (val - mean) / std).collect()
 }
 
-pub fn l2_norm(x: &[Float]) -> Float {
-    x.iter().map(|val| val.powi(2)).sum::<Float>().sqrt()
+pub fn l2_norm(x: &[f64]) -> f64 {
+    x.iter().map(|val| val.powi(2)).sum::<f64>().sqrt()
 }
